Se arreglan cosas en la pantalla de ABM de usuario

+++ FrbaHotel/FrbaHotel/Usuario.cs
@@ -10,6 +10,6 @@ namespace FrbaHotel
     {
         public string id { get; set; }
         public Rol rol { get; set; }
-        public long hotel { get; set; }
+        public int hotel { get; set; }
     }
 }
